requirements.txt
altair==5.3.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
fonttools==4.53.1
gcloud==0.18.3
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.32.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gspread==6.1.2
httplib2==0.22.0
idna==3.7
imageio==2.34.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
networkx==3.3
numba==0.60.0
numpy==2.0.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.27.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-jwt==4.1.0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
resampy==0.4.3
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
scikit-image==0.24.0
scikit-maad==1.4.2
scipy==1.14.0
setuptools==70.3.0
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
streamlit==1.36.0
tenacity==8.5.0
tifffile==2024.7.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
